v0.88.a: Setting and Socket changes.
This commit is a non c++x11 backport of the
reworked Settings and Socket code.

This code is much stricter with paths.
Now all the paths are auto-built on first use.

OTSocket has been re-worked to contain its own settings code,
(OTLog is only for logging now).

The TransportCallback now makes use of <functional> notation.
It is no-longer static. (Yay).